pre-release v8

+++ include/ir.h
@@ -6,16 +6,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "instruction.h"
-#include "opcode.h"
-#include "immediate.h"
-#include "njvm.h"
-#include "pc.h"
 #include "bytecode.h"
+#include "immediate.h"
+#include "instruction.h"
+#include "njvm.h"
+#include "opcode.h"
+#include "pc.h"
 #include "utils.h"
 
-
-void initialize_ir(size_t instruction_count);
+void initialize_ir(unsigned instruction_count);
 void register_instruction(Opcode opcode, Immediate immediate);
 void free_ir(void);
 void print_ir(void);
